+++ Main_poet.py
@@ -5,7 +5,7 @@
 
 from Parameters import Configuration
 from Algorithms.POET.Mutation import mutate_envs
-from Algorithms.POET.LocalTraining import ES_Step, NSGAII_step
+from Algorithms.POET.LocalTraining import Local_Algorithm
 from Algorithms.POET.Transfer import Evaluate_Candidates
 from Utils.Loader import resume_from_folder, prepare_folder
 from Utils.Stats import bundle_stats, append_stats
@@ -112,8 +112,7 @@ for t in range(start_from, args.T):
         E, theta = EA_List[m]
         if args.verbose > 0:
             print(f"\n\t{m} : ", end="", flush=True)
-        theta, threshold[m, t % 5] = ES_Step(theta, E, args, allow_verbose=1)
-        # theta, threshold[m, t % 5] = ES_Step(theta, E, args, allow_verbose=1)
+        theta, threshold[m, t % 5] = Local_Algorithm(theta, E, args, allow_verbose=1)
         EA_List[m] = (E, theta)
 
     # Transfer  -------------------------------------------------------
@@ -139,10 +138,15 @@ for t in range(start_from, args.T):
     print(" Done.")
 
     # Save current execution ------------------------------------------
-    if args.save_mode == "last" and t > 0:
-        os.remove(f'{args.save_to}/Iteration_{t-1}.pickle')
-    with open(f'{args.save_to}/Iteration_{t}.pickle', 'wb') as f:
-        pickle.dump(EA_List, f)
+    if args.save_mode.is_digit():
+        if t % int(args.save_mode) == 0:
+            with open(f'{args.save_to}/Iteration_{t}.pickle', 'wb') as f:
+                pickle.dump(EA_List, f)
+    else:
+        if args.save_mode == "last" and t > 0:
+            os.remove(f'{args.save_to}/Iteration_{t - 1}.pickle')
+        with open(f'{args.save_to}/Iteration_{t}.pickle', 'wb') as f:
+            pickle.dump(EA_List, f)
     with open(f'{args.save_to}/Archive.pickle', 'wb') as f:
         pickle.dump(Configuration.archive, f)
     with open(f"{args.save_to}/TotalBudget.json", 'w') as f:
